requirements.txt
flask
flask-socketio
google-genai
python-dotenv
simple-websocket
